pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["Арсений <90235573+Ribakovan@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.104.0"}
pyjwt = {extras = ["crypto"], version = "^2.8.0"}
python-dotenv = "^1.0.0"
bcrypt = "^4.1.1"


[tool.poetry.group.dev.dependencies]
black = "^23.10.0"
pytest = "^7.4.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
